fix: align upload paths and refactor user management
